skills/sora-watermark-lite/scripts/requirements.txt
numpy>=1.24
ffmpeg-python>=0.2.0
